# docs/openapi-doc.yml
openapi: 3.1.0
info:
    title: Firepit API
    version: 1.7.0
    description: |
        Firepit is a realtime communication platform built with Next.js 16, React 19, Bun, and Appwrite.

        This specification covers the public and first-party client API surface currently used by the application.
        Internal debug and test-only endpoints are intentionally omitted.
    contact:
        name: Firepit
        url: https://github.com/acarlson33/firepit
    license:
        name: GNU General Public License v3.0
        url: https://opensource.org/licenses/GPL-3.0
servers:
    - url: http://localhost:3000
      description: Local development server
    - url: https://firepit.example.com
      description: Example production server
...
components:
    securitySchemes:
        sessionCookie:
            type: apiKey
            in: cookie
            name: a_session_*
            description: Appwrite session cookie used by authenticated first-party clients
    schemas:
        Error:
            type: object
            required:
                - error
            properties:
                error:
                    type: string
                details:
                    description: Optional structured error context
                    oneOf:
                        - type: string
                        - type: object
                          additionalProperties: true
        Success:
            type: object
            properties:
                success:
                    type: boolean
                    default: true
...
        UserProfile:
            type: object
            properties:
                userId:
                    type: string
                displayName:
                    type: string
                userName:
                    type: string
                avatarUrl:
                    type: string
                    format: uri
                pronouns:
                    type: string
                bio:
                    type: string
                profileBackgroundColor:
                    type: string
                    description: Hex color used for the user's profile background when no gradient/image is set.
                    pattern: "^#[0-9A-Fa-f]{6}$"
                    example: "#1a1a2e"
                profileBackgroundGradient:
                    type: string
                    description: CSS gradient string used for the profile background when selected. The server sanitizes this value and only accepts linear-gradient, radial-gradient, or conic-gradient function syntax.
                    pattern: '^(?!.*(?:url\s*\(|data:|javascript:))(linear-gradient|radial-gradient|conic-gradient)\([^;{}<>`\\]+\)$'
                    example: "linear-gradient(135deg, #667eea 0%, #764ba2 100%)"
                profileBackgroundUrl:
                    type: string
                    format: uri
                    description: Resource URI for the uploaded profile background image.
                avatarFramePreset:
                    type: string
                    description: Preset avatar frame ID (for example default-round, seasonal-winter-2026).
                    example: "default-round"
                avatarFrameUrl:
                    type: string
                    format: uri
                    description: Resource URI for the resolved avatar frame image asset.
...
paths:
...
    /api/upload-image:
        post:
            tags: [Uploads]
            summary: Upload an image for message use
            security:
                - sessionCookie: []
            requestBody:
                required: true
                content:
                    multipart/form-data:
                        schema:
                            type: object
                            properties:
                                file:
                                    type: string
                                    format: binary
            responses:
                "200":
                    description: Image uploaded
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    fileId:
                                        type: string
                                    fileUrl:
                                        type: string
                                        format: uri
                                required: [fileId, fileUrl]
                "400":
                    description: Invalid upload request (missing file, unsupported type, or malformed body)
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/Error"
                "401":
                    description: Unauthorized
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/Error"
                "413":
                    description: File too large
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/Error"
        delete:
            tags: [Uploads]
            summary: Delete an uploaded image
            security:
                - sessionCookie: []
            parameters:
                - name: fileId
                  in: query
                  required: true
                  schema:
                      type: string
            responses:
                "200":
                    description: Image deleted
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/Success"
                "403":
                    description: Forbidden — caller does not own the file
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/Error"
                "404":
                    description: Image not found
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/Error"
...
    /api/users/{userId}/profile:
        get:
            tags: [Users]
            summary: Get enriched profile data by user identifier
            security:
                - sessionCookie: []
            parameters:
                - name: userId
                  in: path
                  required: true
                  schema:
                      type: string
            responses:
                "200":
                    description: Enriched user profile
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/UserProfile"
...